requirements.txt
aiohttp
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==3.3.1
certifi==2023.5.7
cffi
charset-normalizer==3.1.0
colorama==0.4.6
cryptography
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-bootstrap4==23.1
django-ckeditor==6.5.1
django-environ==0.10.0
django-js-asset==2.0.0
docker==6.1.3
environ==1.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.7
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
webencodings==0.5.1
websocket-client==1.5.3
yarl==1.9.2
psycopg2==2.9.6
python-dotenv==1.0.0